pyproject.toml
[project]
name = "cf-essay-writer-agent"
version = "0.1.0"
description = "JupyterLab on Tanzu Platform for Cloud Foundry with uv"
requires-python = ">=3.11"
dependencies = [
    "jupyterlab==4.2.5",
    "jupyterlab-git==0.51.2",
    "notebook",
    "ipykernel",
    "python-dotenv",
    "ipywidgets",
    "requests",
    "openai",
    "gradio",
    "langchain",
    "langchain-openai",
    "langchain-experimental",
    "langchain-chroma",
    "langchain[docarray]",
    "datasets",
    "matplotlib",
    "google-generativeai",
    "anthropic",
    "unstructured",
    "pydub",
    "modal",
    "ollama",
    "langgraph-checkpoint-sqlite",
    "tavily-python",
    "langchain-mcp-adapters",
    "matplotlib",
    "unstructured",
    "pydub",
    "modal",
    "langgraph>=0.6.5",
    "grandalf>=0.8",
    "nest-asyncio>=1.6.0",
]

[tool.uv]
dev-dependencies = []